# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SystemGatewayAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Patient/Register": {
      "post": {
        "tags": [
          "Patient"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{token}": {
      "delete": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{Token}": {
      "put": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{Token}/Modules/{ModuleId}": {
      "post": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{Token}/Modules": {
      "post": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{Token}/Modules/{ModuleId}/Version/{VersionId}": {
      "put": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "VersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{token}/Reject/{TherapistEmail}": {
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "TherapistEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{token}/Accept/{therapistEmail}": {
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "therapistEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{token}/Modules": {
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{token}/Modules/{ModuleId}": {
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Patient/{Token}/Therapists": {
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Therapist/{token}/Patients/{email}": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}/Patients/{email}/Modules": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}/Patients/{email}/Modules/{ModuleId}": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{Token}/Patients/{PatientEmail}/Modules/{ModuleId}": {
      "put": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleData": {
            "$ref": "#/components/schemas/ModuleVersion"
          }
        },
        "additionalProperties": false
      },
...
      "Patient": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "profilePicture": {
            "type": "string",
            "nullable": true
          },
          "conditionName": {
            "type": "string",
            "nullable": true
          },
          "conditionAcquisitionDate": {
            "type": "string",
            "format": "date-time"
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            },
            "nullable": true
          },
          "acceptedTherapists": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "requestedTherapists": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}